Update Mutex implementation (#3)
Initially this was going to add a in-memory cache (using sync.map, map[] or similar and mutexes), however I realised that this would likely be more of a "business logic" requirement - this API Library should only handle getting the data. Storing the result of the API Calls (either in memory or to a persistent location (DB, file etc)) should be handled separately depending on requirements of the consumer.

I left the stores that I was looking at implementing inside the "internal" folder for future reference, if something like this is considered again.